requirements.txt
selenium
beautifulsoup4
python-dotenv
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-google-genai>=0.0.5
fastapi>=0.104.1
uvicorn>=0.24.0
chromadb>=0.4.18
tiktoken>=0.5.1
pydantic>=2.4.2
sentence-transformers 
langchain-huggingface